:construction: Fixed data migration

+++ src/store/actions/plants.js
@@ -1,6 +1,6 @@
 import uuid from 'uuid/v4'
 import { blobToBase64String } from 'blob-util'
-import { isBlobbable } from '@/utils/blob'
+import { isBlobbable, getUrlFromBlob } from '@/utils/blob'
 import { iOS } from '@/utils/useragent'
 
 import {
@@ -25,38 +25,62 @@ const namespace = 'plant-'
 const folder = 'plants'
 const fileName = 'cover.png'
 
+async function loadPlantsFirestore (state) {
+  const plants = []
+  const snapshot = await firestoreQuery([['users', state.user.id], [folder]]).get()
+
+  for (const doc of snapshot.docs) {
+    const plant = await firestoreQuery([
+      ['users', state.user.id],
+      [folder, doc.id]
+    ]).get()
+    const plantData = plant.data()
+
+    if (plantData.imageURL) {
+      plantData.imageURL = await downloadFile(plantData.imageURL)
+
+      if (state.storage.migrationMode) {
+        const photo = await fetch(plantData.imageURL)
+        plantData.blob = await photo.blob()
+        plantData.imageURL = getUrlFromBlob(plantData.blob)
+      }
+    }
+
+    plants.push(plantData)
+  }
+
+  return plants
+}
+
+async function loadPlantsLocalforage () {
+  const plants = []
+  const values = await getEntryLF(namespace)
+    .then(plantData => {
+      const copy = plantData
+      delete copy[namespace + 'undefined']
+      return copy
+    })
+    .then(Object.values)
+
+  for (const item of values) {
+    plants.push(item)
+  }
+
+  return plants
+}
+
 export async function loadPlants ({ state, commit }) {
   commit('LOAD_PLANTS_PROGRESS')
   let plants = []
 
   if (state.storage.type === 'cloud' && state.user.id) {
     try {
-      const snapshot = await firestoreQuery([['users', state.user.id], [folder]]).get()
-      for (const doc of snapshot.docs) {
-        const plant = await firestoreQuery([
-          ['users', state.user.id],
-          [folder, doc.id]
-        ]).get()
-        const plantData = plant.data()
-        plants.push({
-          ...plantData,
-          imageURL: plantData.imageURL && await downloadFile(plantData.imageURL)
-        })
-      }
+      plants = await loadPlantsFirestore(state)
     } catch (error) {
       commit('LOAD_PLANTS_FAILURE')
     }
    } else {
-    const values = await getEntryLF(namespace)
-      .then(plantData => {
-        const copy = plantData
-        delete copy[namespace + 'undefined']
-        return copy
-      })
-      .then(Object.values)
-    for (const item of values) {
-      plants.push((item))
-    }
+    plants = await loadPlantsLocalforage()
   }
 
   commit('LOAD_PLANTS_SUCCESS', { plants })
@@ -66,9 +90,37 @@ export function loadPlantItem ({ state, commit }, guid) {
   commit('LOAD_PLANT_ITEM', { guid })
 }
 
+async function addPlantFirestore (state, meta) {
+  const path = [['users', state.user.id], [folder, meta.guid]]
+  const hasFile = meta.blob && isBlobbable(meta.blob)
+  const blob = meta.blob
+  await addEntryFire(path, {
+    ...meta,
+    blob: null,
+    imageURL: hasFile && `${storagePath(path)}/${fileName}`
+  })
+  if (hasFile) {
+    await uploadFile(path.concat(fileName), blob)
+  }
+}
+
+async function addPlantLocalforage (state, meta) {
+  // FIXME: This is generally a bad idea. Use feature detection instead.
+  // However, I could not find a reliable way to test if IndexedDB supports blobs,
+  // as it fails silently. We have to convert the blob to base64,
+  // because mobile Safari 10 has a bug with storing Blobs in IndexedDB.
+
+  if (iOS && !!meta.blob) {
+    const base64String = await blobToBase64String(meta.blob)
+    await addEntryLF(namespace + meta.guid, { ...meta, blob: base64String })
+  } else {
+    await addEntryLF(namespace + meta.guid, meta)
+  }
+}
+
 export async function addPlant ({ state, commit }, data) {
   commit('ADD_PLANT_PROGRESS')
-  let meta = state.storage.migrationMode ? data : {
+  const meta = state.storage.migrationMode ? data : {
     ...data,
     guid: uuid(),
     created: Date.now(),
@@ -76,44 +128,20 @@ export async function addPlant ({ state, commit }, data) {
   }
 
   if (state.storage.type === 'cloud') {
-    const path = [['users', state.user.id], [folder, meta.guid]]
-    const hasFile = meta.blob && isBlobbable(meta.blob)
-    await addEntryFire(path, {
-      ...meta,
-      blob: null,
-      imageURL: hasFile && `${storagePath(path)}/${fileName}`
-    })
-    if (hasFile) {
-      await uploadFile(path.concat(fileName), meta.blob)
+    try {
+      await addPlantFirestore(state, meta)
+    } catch (error) {
+      commit('ADD_PLANT_FAILURE')
     }
   } else {
-    // If we're in migration mode, we want to create a blob from the image
-    if (state.storage.migrationMode) {
-      const image = await fetch(meta.imageURL)
-      meta.blob = image.blob()
-      meta.imageURL = null
-    }
-
-    // FIXME: This is generally a bad idea. Use feature detection instead.
-    // However, I could not find a reliable way to test if IndexedDB supports blobs,
-    // as it fails silently. We have to convert the blob to base64,
-    // because mobile Safari 10 has a bug with storing Blobs in IndexedDB.
-    if (iOS && !!meta.blob) {
-      // 1. Turn blob into base64 string (only needed for storage)
-      const base64String = await blobToBase64String(meta.blob)
-      const config = Object.assign({}, meta, { blob: base64String })
-
-      await addEntryLF(namespace + config.guid, config)
-      meta = Object.assign({}, config, { blob: meta.blob })
-    } else {
-      await addEntryLF(namespace + meta.guid, meta)
-    }
+    await addPlantLocalforage(state, meta)
   }
 
   if (!state.storage.migrationMode) {
-    commit('ADD_PLANT', { item: meta })
-    return meta.guid
+    commit('ADD_PLANT_SUCCESS', { item: meta })
   }
+
+  return meta.guid
 }
 
 export async function updatePlant (action, { state, commit }, data) {
@@ -140,16 +168,22 @@ export async function deletePlants ({ state, commit }, items) {
   commit('DELETE_PLANT_PROGRESS')
 
   if (state.storage.type === 'cloud') {
-    await Promise.all(items.map(async item => {
-      const path = [['users', state.user.id], [folder, item.guid]]
-      await deleteEntryFire(path)
-      await deleteFile(path.concat(fileName))
-    }))
+    try {
+      await Promise.all(items.map(async item => {
+        const path = [['users', state.user.id], [folder, item.guid]]
+        await deleteEntryFire(path)
+        if (item.blob && isBlobbable(item.blob)) {
+          await deleteFile(path.concat(fileName))
+        }
+      }))
+    } catch (error) {
+      commit('DELETE_PLANTS_FAILURE')
+    }
   } else {
     await Promise.all(items.map(item => deleteEntryLF(namespace + item.guid, item)))
   }
 
   if (!state.storage.migrationMode) {
-    commit('DELETE_PLANTS', { items })
+    commit('DELETE_PLANTS_SUCCESS', { items })
   }
 }
